# docs/api/openapi_v1.yaml
openapi: 3.0.3
info:
  title: Monitor Legislativo v4 API
  description: |
    Comprehensive API for monitoring Brazilian legislative documents and regulatory changes.
    
    ## Features
    - **Document Search**: Advanced search with filters and facets
    - **Real-time Monitoring**: Set up alerts for legislative changes
    - **Smart Alerts**: ML-powered relevance scoring
    - **Export Capabilities**: Multiple format support
    - **User Management**: Authentication and authorization
    
    ## Authentication
    This API uses JWT (JSON Web Token) authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to ensure fair usage:
    - **Authenticated users**: 1000 requests/hour
    - **Search endpoints**: 100 requests/minute
    - **Export endpoints**: 10 requests/minute
    
    ## Error Handling
    All endpoints return standardized error responses with structured error information.
    
  version: 1.0.0
  contact:
    name: Monitor Legislativo Support
    email: support@monitorlegislativo.com
    url: https://monitorlegislativo.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.monitorlegislativo.com/v1
    description: Production server
  - url: https://staging-api.monitorlegislativo.com/v1
    description: Staging server
  - url: http://localhost:5000/api/v1
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              standard_user:
                summary: Standard user login
                value:
                  username: "user@example.com"
                  password: "securepassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      description: Get new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    StandardResponse:
      type: object
      required:
        - status
        - data
        - metadata
      properties:
        status:
          type: string
          enum: [success, error, warning, partial]
          description: Response status
        data:
          description: Response data (type varies by endpoint)
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          description: Error details (present when status is error)
        warnings:
          type: array
          items:
            type: string
          description: Warning messages
    ResponseMetadata:
      type: object
      required:
        - timestamp
        - version
        - request_id
        - execution_time_ms
      properties:
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        version:
          type: string
          description: API version
          example: "v1"
        request_id:
          type: string
          format: uuid
          description: Unique request identifier
        execution_time_ms:
          type: number
          format: float
          description: Request execution time in milliseconds
        total_count:
          type: integer
          description: Total number of items (for paginated responses)
        page:
          type: integer
          description: Current page number
        per_page:
          type: integer
          description: Items per page
        has_more:
          type: boolean
          description: Whether more pages are available
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Query parameter is required"
        details:
          type: object
          description: Additional error context

    # Authentication Schemas
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          format: email
          description: User email address
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          description: User password
          example: "securepassword123"
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                  description: JWT access token
                refresh_token:
                  type: string
                  description: JWT refresh token
                token_type:
                  type: string
                  enum: [Bearer]
                expires_in:
                  type: integer
                  description: Token expiration time in seconds
                user:
                  $ref: '#/components/schemas/User'
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Valid refresh token
    TokenResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                token_type:
                  type: string
                  enum: [Bearer]
                expires_in:
                  type: integer
    User:
      type: object
      properties:
        id:
          type: integer
          description: User ID
        username:
          type: string
          description: Username
        email:
          type: string
          format: email
          description: Email address
        first_name:
          type: string
          description: First name
        last_name:
          type: string
          description: Last name
        role:
          type: string
          enum: [admin, analyst, user, readonly]
          description: User role

    # Document Schemas
...
  responses:
...
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/StandardResponse'
              - type: object
                properties:
                  status:
                    type: string
                    enum: [error]
          example:
            status: error
            data: null
            metadata:
              timestamp: "2025-01-06T12:00:00Z"
              version: "v1"
              request_id: "123e4567-e89b-12d3-a456-426614174000"
              execution_time_ms: 12.1
            errors:
              - code: "AUTHENTICATION_ERROR"
                message: "Authentication required"
...
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/StandardResponse'
              - type: object
                properties:
                  status:
                    type: string
                    enum: [error]
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per time window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Time when rate limit resets (Unix timestamp)
...